Feature/stock db

+++ src/main/java/com/chan/stock_portfolio_backtest_api/domain/CalcIndexPrice.java
@@ -19,8 +19,8 @@ public class CalcIndexPrice {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Integer id;
 
-    private Integer price;
-    private float monthlyRor;
+    private Float price;
+    private Float monthlyRor;
     private LocalDate baseDate;
 
     @JsonIgnore
